Implement history saving and loading: * modify HistoryItem to use SoldItems directly; * add annotations to HistoryItem; * modify POSdb scheme to support modified HistoryItems * add HistoryItem to DB mappings; * modify PaymentWindow logic to support saving history via SalesDomainController; * modify SalesDomainController and implementation to support loading history; * modify SalesSystemModel to automatically populate history from DB on startup; * add default constructors to HistoryItem and SoldItem to support loading from DB; * minor changes in code to allow loading history on startup.
Fix typos in SoldItem.

+++ trunk/src/ee/ut/math/tvt/salessystem/ui/model/PurchaseInfoTableModel.java
@@ -82,7 +82,7 @@ public class PurchaseInfoTableModel extends SalesSystemTableModel<SoldItem> {
 	 */
 	public double getTotalSum() {
 
-		int totalSum = 0;
+		double totalSum = 0;
 		
 		Iterator<SoldItem> it = rows.iterator();
